+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `FullyQualifiedImportCheck` analysis rule, which flags non-fully qualified imports.
+- **API:** `UnitImportNameDeclaration::isAlias` method.
+
 ### Changed
 
 - Non-trivial `inherited` expressions are excluded in `RedundantParentheses`.
@@ -19,7 +24,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- `FullyQualifiedImportCheck` analysis rule, which flags non-fully qualified imports.
 - **API:** `CaseItemStatementNode::getExpressions` method.
 
 ### Changed